Pass a bunch of optional arguments to model_dump and python_dump through (#38)
* Pass a bunch of optional arguments to model_dump and python_dump through is specified.

* lint

* Use PEP-692 **dump_args: Unpack[PydanticDumpArgs] for long list of optionaly pydantic model dump arguments for bretvity, clarity and safety.

This required bumping python version to 3.12 but since we don't have legacy deployments, we should be fine.

Added unit tests for the optional arguments and fixed some test names.

* Use PEP-692 **dump_args: Unpack[PydanticDumpArgs] for long list of optionaly pydantic model dump arguments for bretvity, clarity and safety.

This required bumping python version to 3.12 but since we don't have legacy deployments, we should be fine.

Added unit tests for the optional arguments and fixed some test names.

* Newer type alias and generics syntax

* Small changes fix

---------

Co-authored-by: Logi <logi.ragnarsson@ankeri.is>